client/client_update: split update client and http client setup
Move deployments service client (aka. update client) to separate
client_udpate.go file. Leave a simplified HTTP client setup in
client.go.
